+++ packages/kit/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sveltejs/kit
 
+## 0.0.28
+
+### Patch Changes
+
+- Add svelte-kit start command
+
 ## 0.0.27
 
 ### Patch Changes